Add extra step in incompatible investments test
The test was always passing because incompatible investments are only
shown after clicking on the "Show all" link. So now we're clicking the
link  before checking the "incompatibles" table isn't shown.

I don't think "incompatibles" is a valid English word; for now, I'm
leaving it like this until a native English speaker reviews it.