Add Rails/RedundantTravelBack rubocop rule
This rule was added in rubocop-rails 2.12.0. It doesn't catch the case
we've seen the most in the past, though: using `travel_back` to finish
the test will not raise an offense.

However, it does detect a useless `travel_back` call in `after` blocks,
so I guess it's better than nothing.