Update remote translation docs
Due to the change of the translator-text gem to the bing_translator gem,
the documentation has been reviewed and changes have been detected
in the links, prices and azure administration panel.

We removed the Google and DeepL prices, since we will not be reviewing
the prices of the different providers and they may be out of date.