Explicitly enable ERB linters
I thought SelfClosingTag was enabled since it was on the list and the
`enabled: false` got lost among all the `enabled: true`. ERB Lint 0.1
and later will also add RequireInputAutocomplete as a default linter,
and we're not interested in it, at least for now.

So, just like we do for Rubocop, we're disabling all linters and
enabling the ones we use explicitly.