Bump `leaflet` from `1.5.1` to `1.9.3`
Newer versions have more capabilities and also fix lot of errors,
usability and accessibility issues.

Note that we're using an external gem to rails-assets.org as it's
not supports the rails-assets pipeline. This gems wraps the
original one and make the code work with the default rails assets
pipeline as usual.