Remove unused code for legislation proposals votes
In these commits 38ba5e159b and 8805037e2fad we added the parameter
"vote_url" in the call to the partial "votes" in collaborative legislation
proposals.

It seems that this parameter is only used in the proposals module and not in
collaborative legislation proposals.

While it is true that in the partial "votes" in these commits this parameter
"vote_url" is referred to, in commit 276baedcf it seems to be removed.

For this reason, we remove the parameter.