Remove unnecessary code in proposal scope
The `id:` condition works with a list of proposals that might include
nil items, so there's no need to compact it and manually get the IDs.

Note this scope is probably inefficient, since it instantiates proposal
objects (with the `find_voted_items` method) for something that could be
done with a database query. We might change it in he future.